Cleanup Swords & Taming commands

+++ src/main/java/com/gmail/nossr50/config/AdvancedConfig.java
@@ -136,12 +136,12 @@ public class AdvancedConfig extends ConfigLoader {
     public int getArcaneForgingRankLevels4() { return config.getInt("Skills.Repair.Arcane_Forging.Rank_Levels.Rank_4", 750); }
 
     /* SWORDS */
-    public int getBleedChanceMax() { return config.getInt("Skills.Swords.Bleed_ChanceMax", 75); }
+    public double getBleedChanceMax() { return config.getDouble("Skills.Swords.Bleed_ChanceMax", 75.0D); }
     public int getBleedMaxBonusLevel() { return config.getInt("Skills.Swords.Bleed_MaxBonusLevel", 750); }
     public int getBleedMaxTicks() { return config.getInt("Skills.Swords.Bleed_MaxTicks", 3); }
     public int getBleedBaseTicks() { return config.getInt("Skills.Swords.Bleed_BaseTicks", 2); }
 
-    public int getCounterChanceMax() { return config.getInt("Skills.Swords.Counter_ChanceMax", 30); }
+    public double getCounterChanceMax() { return config.getDouble("Skills.Swords.Counter_ChanceMax", 30.0D); }
     public int getCounterMaxBonusLevel() { return config.getInt("Skills.Swords.Counter_MaxBonusLevel", 600); }
     public int getCounterModifier() { return config.getInt("Skills.Swords.Counter_DamageModifier", 2); }
 
@@ -149,13 +149,13 @@ public class AdvancedConfig extends ConfigLoader {
     public int getSerratedStrikesTicks() { return config.getInt("Skills.Swords.SerratedStrikes_BleedTicks", 5); }
 
     /* TAMING */
-    public int getGoreChanceMax() { return config.getInt("Skills.Taming.Gore_ChanceMax", 100); }
+    public double getGoreChanceMax() { return config.getInt("Skills.Taming.Gore_ChanceMax", 100); }
     public int getGoreMaxBonusLevel() { return config.getInt("Skills.Taming.Gore_MaxBonusLevel", 1000); }
     public int getGoreBleedTicks() { return config.getInt("Skills.Taming.Gore_BleedTicks", 2); }
     public int getGoreModifier() { return config.getInt("Skills.Taming.Gore_Modifier", 2); }
 
     public int getFastFoodUnlock() { return config.getInt("Skills.Taming.FastFood_UnlockLevel", 50); }
-    public int getFastFoodChance() { return config.getInt("Skills.Taming.FastFood_Chance", 50); }
+    public double getFastFoodChance() { return config.getInt("Skills.Taming.FastFood_Chance", 50); }
 
     public int getEnviromentallyAwareUnlock() { return config.getInt("Skills.Taming.EnvironmentallyAware_UnlockLevel", 100); }
